requirements.txt
py-cord == 2.2.2
discord.py == 2.4.0
python-dotenv == 1.1.0
google-genai == 1.12.1
